chore: Update eslint config to use @ysc instead of @repo
This commit updates the eslint configuration files to use the @ysc organization instead of the @repo organization. It modifies the .eslintrc.js file in the root directory, as well as the .eslintrc.cjs files in the apps/docs and apps/web directories. The "extends" property is updated to use the @ysc/eslint-config package instead of the @repo/eslint-config package.

Note: This suggested commit message follows the convention used in the recent repository commits for this project.